Split search sort into component
